Fix audio input for LTC sync on Windows
Use 16 bits input instead of 8 bits (8-bits setting always returns
samples whose value is zero on Windows).

Ask for a reasonable latency (Asking for zero latency enables an
internal buffering in portaudio at the level of 2 or 3 samples,
and leads to permanent overflow), and read the data in 512-samples
buffers.

In the long run, ideally, decoding the LTC data inside the PortAudio
callback should be avoided to prevent audio overflows. Instead the
callback should simply push the data to a ringbuffer, which should
be decoded in another thread.