(LtcReader) Fix timecode type
Prior to this commit, the LTC decoder converts the timecode to a frame
number with a hardcoded 25 fps timecode type. PhLTCReader clock is
indeed set to 25 fps at creation, however it is changed to the video
fps in Synchronizer::onVideoTCTypeChanged. With a 24 fps video, the
LTC decoder would slowly drift.

To handle this situation, use the clock timecode type instead of
hardcoding it at 25 fps.