Enable bilinear filtering for nicer video scaling.
Prior to this commit, the video is scaled with a nearest-neighbour
filter. This produces 'blocky' images on screen. By switching to
bilinear filtering (which is standard), much nicer images are produced.

The change to bilinear scaling implies an additional modification to
specify how the filter handles the pixels at the edge of the video.
The scaling by default uses a `repeat' mode, which is only
appropriate when drawing a repeating pattern. So we use `clamp to
edge' as appropriate here for all textures apart from the strip
background.