Switch from QGLWidget to a QQuickWidget
With a QQuickWidget, the OpenGL drawings are done in the slot
connected to beforeRendering() signals.

There may be a slight performance decrease since the OpenGL
drawings are done to a framebuffer.

The immediate gain is that QML can be used to draw instead of raw
OpenGL ! QML is much more friendly and powerful, so nicer
user-interfaces are possible.
The underlying scenegraph is designed for high-performance, so that
complicated drawings (gradients, etc.) are achievable at high
framerates.

Fix the refresh rate

The QGLWidget used before was v-synced, so the effective refresh
rate was the screen refresh rate, even if the code did set it to be
twice as fast.

With the switch to a QQuickWidget, the v-sync is turned off (it can
only be kept on with a pure QWidget or a pure QML app).

Fix the event filter

The installation was removed by mistake before ?

Fix the media panel fade in with the new widget