Switch from QGLWidget to a QQuickWidget
With a QQuickWidget, the OpenGL drawings are done in the slot
connected to beforeRendering() signals.

There may be a slight performance decrease since the OpenGL
drawings are done to a framebuffer.

The immediate gain is that QML can be used to draw instead of raw
OpenGL ! QML is much more friendly and powerful, so nicer
user-interfaces are possible.
The underlying scenegraph is designed for high-performance, so that
complicated drawings (gradients, etc.) are achievable at high
framerates.