Draw the strip texts with QML
Here the strategy changes a lot from the previous OpenGL code...
The text is rendered and scaled at painting time ! This may need
some optimisation work to make it fast enough. But the immediate
benefit is the internationalization since UTF-8 is supported.

Custom font and color are not yet supported.:wq