(MidiTool) Add a specific setting for the existing port use
This decouples the logic for the 2 settings midiInputPortName and
virtualInputPortName, so that it will be easier to disable the virtual
feature on Windows where it is not available.

It also prevents issues when a port disappears between two Joker
uses (the MIDI interface is unplugged for example). Before this commit,
the old interface name would be used as the virtual port name...