Performance improvement in the video engine
In the current design, there is a significant performance bottleneck in
maintaining the _requestedFrames list each time the clock advances.

Here this bottleneck is removed by making the decoder class work
differently. Instead of maintaining a list of requested frames, the
decoder compares the current time, the readahead time, and the playback
direction to decide whether to continue decoding the next frame, to seek
or stop.

Additionally, the seeking threshold is reduced to a minimum for MJPEG
videos where every frames are keyframes.