Travis: bypass keys and coverity scan in PR builds
PR builds do not have access to secure variables from .travis.yml.

As it is, this prevents PR builds from building Joker and its test suite
in the OSX environment, because the script to add keys cannot succeed.

Additionnally, the Coverity scan script fails too because the token is
not available.

In this commit, those scripts are not run if the build is coming from
a Pull Request.